+++ app/Api/Drills.ts
@@ -1,6 +1,6 @@
 import { calculateNotes, sudokuStrategy, sudokuStrategyArray } from "sudokuru";
-import { puzzle } from "../Types/Puzzle.Types";
-import { drill } from "../Types/Puzzle.Types";
+import { puzzle } from "./Puzzle.Types";
+import { drill } from "./Puzzle.Types";
 import { returnLocalDrillGame } from "../Functions/LocalDatabase";
 
 /**
